first sketch in player.py
sketched minimax, havent tested, also still havent covered the case where there are no children (but will that be necessary considering the character of the terminal states? wont children always be created? havent looked into the provided code deeply enough)

code could be moved into an object in a module as suggested by the instructions but thats for laterz if at all